Mvbcra

+++ src/components/card.tsx
@@ -1,6 +1,5 @@
-import * as React from "react";
+import React, { useCallback } from "react";
 import distance from "date-fns/formatDistanceToNow";
-import parseISO from "date-fns/parseISO";
 import { EditLink, PreviewLink } from "./entry-links";
 import { IPartialFeedItem } from "../reducers/dashboard";
 
@@ -9,14 +8,16 @@ export interface ICardProps {
   excerpt?: string;
   id: string;
   dateAdded: Date | string;
-  onDelete?: ({ id }: Partial<IPartialFeedItem>) => void;
+  onDelete: ({ id, title }: IPartialFeedItem) => void;
   public: boolean;
 }
 
 export default function Card(props: ICardProps): JSX.Element {
-  function onDelete() {
-    props.onDelete({ id: props.id, title: props.title });
-  }
+  const onDelete = useCallback(() => {
+    if (props.onDelete) {
+      props.onDelete({ id: props.id, title: props.title });
+    }
+  }, [props.onDelete, props.id, props.title]);
 
   return (
     <div className="Sheet Card" data-testid="CARD">
@@ -26,7 +27,7 @@ export default function Card(props: ICardProps): JSX.Element {
         </h2>
         {props.dateAdded && (
           <small className="CardMeta">
-            added {distance(parseISO(props.dateAdded.toString()))} ago
+            added {distance(new Date(props.dateAdded))} ago
           </small>
         )}
       </header>
